% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{llike}
\alias{llike}
\title{Log-Likelihood for ordinal regression}
\usage{
llike(Y, W, tau)
}
\arguments{
\item{Y}{ordinal response}

\item{W}{global variance}

\item{tau}{cut-points}
}
\value{
value of log-likelihood at Y, W and tau
}
\description{
This function evaluates the log-likelihood
of the response for a given value of the parameters.
}
\examples{
set.seed(1234)
K = 3 # number of response categories
Y = sample(1:K, 10, replace=TRUE) # generate responses
W = 1 
tau = c(-Inf, -1, 1, Inf) # set parameter values
llike(Y, W, tau)

}
