% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2D2ordinal-package.R
\docType{package}
\name{R2D2ordinal-package}
\alias{R2D2ordinal}
\alias{R2D2ordinal-package}
\title{The 'R2D2ordinal' package.}
\description{
This package implements the methods from "Pseudo-R2D2 prior for high-dimensional ordinal regression" by Eric Yanchenko.
}
\author{
\strong{Maintainer}: Eric Yanchenko \email{eyanchenko@aiu.ac.jp}

}
\keyword{internal}
